requirements.txt
# Enhanced Development Playground Requirements
# ============================================

# Core AGNO framework
agno>=0.1.0

# FastAPI and web server
fastapi>=0.104.1
uvicorn[standard]>=0.24.0

# OpenAI integration
openai>=1.0.0

# Environment and configuration
python-dotenv>=1.0.0

# Database and persistence
sqlalchemy>=2.0.0
sqlite3  # Built-in with Python

# Data validation and parsing
pydantic>=2.0.0

# MongoDB support (for generated backends)
motor>=3.3.0
pymongo>=4.6.0

# Authentication (for generated backends)
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.6

# Development and testing
pytest>=7.0.0
pytest-asyncio>=0.21.0

# Utilities
requests>=2.31.0
aiofiles>=23.0.0

# Optional: Enhanced logging
rich>=13.0.0
colorama>=0.4.6

# Optional: Performance monitoring
psutil>=5.9.0